Refactors the testbed
